Allow more lax env selection from computed factors
The previous env selection was rather strict when there were more than 2
factors supplied and this change allows users to relax it with the
section value "envs_are_optional" which will become the default in a
future release.

Given the following factors:

  [{"py38", "lint"}, {"reqV1", "reqV2"}, {"opReqV1", "opReqV2"}]

The existing env selection would only match:

- py38-reqV1-opReqV1
- py38-reqV2-opReqV1
- py38-reqV1-opReqV2
- py38-reqV2-opReqV2
  ...
- lint-reqV1-opReqV1
  ...

It would fail to match the single factor "lint". Although, this may be
correct for required factors, but for something like "lint" it may not
need the additional factors that are required with the previous env
selection.

This change allows selecting the following envs:

- py38
- py38-reqV1
- py38-reqV2
- py38-opReqV1
- py38-opReqV2
- py38-reqV1-opReqV1
- py38-reqV2-opReqV1
- py38-reqV2-opReqV2
- lint
- ...

In addition, this change makes the order of the factors no longer important:

- py38-opReqV1-reqV1
- reqV1-opReqV1-py38

All of these permutations are bound by the envlist that the user defines
in their tox configuration so it is up to the user to keep their
configuration organized and not go crazy with their factor ordering.